Distribution_Build/spidy_lib/AI_Engine/requirements.txt
langchain
langchain-openai
langchain-google-genai
python-dotenv
pydantic
requests
